remove parallel code that was used for comparison purposes
