+++ flexBPE/CMakeLists.txt
@@ -1,14 +1,12 @@
 # list sourcefiles for convenience
 set(LIB_SRC 
-    flexBPE.cpp
-    flexBPE_trainer.cpp)
+    flexBPE.cpp)
 set(APP_SRC
     main.cpp)
 
 # create library and executable
 add_library(flexbpe SHARED ${LIB_SRC})
 add_executable(flexbpe-bin ${APP_SRC})
-add_executable(flexbpe2-bin "main2.cpp")
 
 # link the executable to the shared library
 set(CMAKE_INSTALL_RPATH "${ORIGIN}")
@@ -16,7 +14,6 @@ set(CMAKE_BUILD_WITH_INSTALL_RPATH TRUE)
 
 # set some properties
 set_target_properties(flexbpe-bin PROPERTIES OUTPUT_NAME "flexbpe")
-set_target_properties(flexbpe2-bin PROPERTIES OUTPUT_NAME "flexbpe2")
 set_target_properties(flexbpe PROPERTIES 
 	                      VERSION ${PROJECT_VERSION} 
                               SOVERSION 0
@@ -41,5 +38,3 @@ find_package(Threads REQUIRED)
 target_link_libraries(flexbpe Threads::Threads)
 # link executable to shared library and threads
 target_link_libraries(flexbpe-bin flexbpe Threads::Threads)
-
-target_link_libraries(flexbpe2-bin flexbpe Threads::Threads)
