+++ flexBPE/flexBPE.cpp
@@ -4,13 +4,14 @@ namespace flexBPE {
 using namespace std;
 
 const size_t kMaxPairs = 1000 * 1000 * 1000;
-const size_t kThreads = max(1, min(10, int(thread::hardware_concurrency())));
-const char *kEndWord = "</w>";
-const size_t kEndWordLength = 4;
-const char *kTokenDelim = "@@";
-const size_t kTokenDelimLength = 2;
 
-int safeOpen(const char *file_path, int flags, mode_t mode = 0) {
+BPETrainer::BPETrainer(const char *jEndWord, const size_t jEndWordLength,
+                       const char *jTokenDelim, const size_t jTokenDelimLength,
+                       const size_t jThreads)
+    : jThreads(jThreads), jEndWord(jEndWord), jEndWordLength(jEndWordLength),
+      jTokenDelim(jTokenDelim), jTokenDelimLength(jTokenDelimLength) {}
+
+int BPETrainer::safeOpen(const char *file_path, int flags, mode_t mode = 0) {
   int fd = open(file_path, flags, mode);
   if (fd < 0) {
     fprintf(stderr, "Cannot open text file %s\n", file_path);
@@ -19,7 +20,8 @@ int safeOpen(const char *file_path, int flags, mode_t mode = 0) {
   return fd;
 }
 
-void readText(const char *fp, unordered_map<string, uint32_t> &word_count) {
+void BPETrainer::readText(const char *fp,
+                          unordered_map<string, uint32_t> &word_count) {
   string cur_word;
   uint64_t total = 0;
   auto deal_with_char = [&](char cur_char) {
@@ -62,8 +64,11 @@ void readText(const char *fp, unordered_map<string, uint32_t> &word_count) {
           word_count.size());
 }
 
-std::pair<size_t, uint64_t> output_or_count(unordered_map<string, string> &bpe,
-                                            size_t size, char *f, char *fo) {
+using ssp = pair<size_t, uint64_t>;
+
+pair<size_t, uint64_t>
+BPETrainer::output_or_count(unordered_map<string, string> &bpe, size_t size,
+                            char *f, char *fo) {
   string cur_word;
   size_t charOut = 0;
   uint64_t total = 0;
@@ -97,9 +102,8 @@ std::pair<size_t, uint64_t> output_or_count(unordered_map<string, string> &bpe,
   return std::make_pair(charOut, total);
 }
 
-void outputText(const char *fpo, const char *fp,
-                unordered_map<string, string> &bpe) {
-
+void BPETrainer::outputText(const char *fpo, const char *fp,
+                            unordered_map<string, string> &bpe) {
   int fd = safeOpen(fp, O_RDONLY);
   auto fdOut = safeOpen(fpo, O_RDWR | O_CREAT | O_TRUNC, 0666);
 
@@ -132,11 +136,11 @@ void outputText(const char *fpo, const char *fp,
   close(fd);
 }
 
-void tokenize(const unordered_map<string, uint32_t> &word_count,
-              unordered_map<string, uint32_t> &token_to_int,
-              vector<string> &int_to_token, vector<list<uint32_t>> &words,
-              vector<int32_t> &counts) {
-
+void BPETrainer::tokenize(const unordered_map<string, uint32_t> &word_count,
+                          unordered_map<string, uint32_t> &token_to_int,
+                          vector<string> &int_to_token,
+                          vector<list<uint32_t>> &words,
+                          vector<int32_t> &counts) {
   for (auto &x : word_count) {
     auto &word = x.first;
 
@@ -161,7 +165,7 @@ void tokenize(const unordered_map<string, uint32_t> &word_count,
       }
       pos++;
     }
-    auto new_token = word.substr(lastStart, string::npos) + kEndWord;
+    auto new_token = word.substr(lastStart, string::npos) + jEndWord;
     if (token_to_int.count(new_token) == 0) {
       int_to_token.push_back(new_token);
       token_to_int[new_token] = int_to_token.size() - 1;
@@ -170,9 +174,8 @@ void tokenize(const unordered_map<string, uint32_t> &word_count,
   }
 }
 
-void tokenize_str(const unordered_map<string, uint32_t> &word_count,
-                  unordered_map<string, vector<string>> &words) {
-
+void BPETrainer::tokenize_str(const unordered_map<string, uint32_t> &word_count,
+                              unordered_map<string, vector<string>> &words) {
   for (auto &x : word_count) {
     auto &word = x.first;
     words[word] = vector<string>();
@@ -190,12 +193,12 @@ void tokenize_str(const unordered_map<string, uint32_t> &word_count,
       }
       pos++;
     }
-    auto new_token = word.substr(lastStart, string::npos) + kEndWord;
+    auto new_token = word.substr(lastStart, string::npos) + jEndWord;
     words[word].push_back(new_token);
   }
 }
 
-void count_in_word(
+void BPETrainer::count_in_word(
     list<uint32_t> &word, uint32_t wi, uint32_t count, pc &pair_counts,
     vector<pair<int32_t, tp>> &contiguous_counts,
     unordered_map<tp, unordered_set<uint32_t>, pair_hash> &where) {
@@ -227,8 +230,8 @@ void count_in_word(
   }
 }
 
-void find_maxp(vector<pair<int32_t, tp>> &contiguous_counts, tp &maxp,
-               int32_t &max_c) {
+void BPETrainer::find_maxp(vector<pair<int32_t, tp>> &contiguous_counts,
+                           tp &maxp, int32_t &max_c) {
   max_c = 0;
   for (auto &x : contiguous_counts) {
     if (x.first > max_c) {
@@ -240,7 +243,7 @@ void find_maxp(vector<pair<int32_t, tp>> &contiguous_counts, tp &maxp,
   }
 }
 
-void getvocab(const char *inputFile1, const char *inputFile2) {
+void BPETrainer::getvocab(const char *inputFile1, const char *inputFile2) {
   // get vocab
   unordered_map<string, uint32_t> word_count;
   readText(inputFile1, word_count);
@@ -262,8 +265,8 @@ void getvocab(const char *inputFile1, const char *inputFile2) {
     cout << element.first << " " << element.second << endl;
 }
 
-void learnbpe(const uint32_t kNPairs, const char *inputFile1,
-              const char *inputFile2) {
+void BPETrainer::learnbpe(const uint32_t kNPairs, const char *inputFile1,
+                          const char *inputFile2) {
   // get vocab
   unordered_map<string, uint32_t> word_count;
   readText(inputFile1, word_count);
@@ -371,7 +374,7 @@ void learnbpe(const uint32_t kNPairs, const char *inputFile1,
   }
 }
 
-void split(vector<string> &splits, const string &text, char sep) {
+void BPETrainer::split(vector<string> &splits, const string &text, char sep) {
   size_t start = 0, end = 0;
   while ((end = text.find(sep, start)) != string::npos) {
     if (end != start)
@@ -382,7 +385,8 @@ void split(vector<string> &splits, const string &text, char sep) {
     splits.push_back(text.substr(start));
 }
 
-void readVocab(const char *fp, unordered_map<string, uint32_t> &vocab) {
+void BPETrainer::readVocab(const char *fp,
+                           unordered_map<string, uint32_t> &voc) {
   ifstream file(fp);
   if (!file) {
     fprintf(stderr, "Cannot open vocabulary file %s\n", fp);
@@ -395,17 +399,18 @@ void readVocab(const char *fp, unordered_map<string, uint32_t> &vocab) {
     vector<string> splits;
     split(splits, line, ' ');
     assert(splits.size() == 2);
-    assert(vocab.find(splits[0]) == vocab.end());
+    assert(voc.find(splits[0]) == voc.end());
     int count = stoi(splits[1]);
-    vocab[splits[0]] = count;
+    voc[splits[0]] = count;
     total += count;
   }
   fprintf(stderr, "Read %lu words (%lu unique) from vocabulary file.\n", total,
-          vocab.size());
+          voc.size());
 }
 
-void readCodes(const char *fp, unordered_map<tps, uint32_t, pair_hash> &codes,
-               unordered_map<string, tps> &reversed_codes) {
+void BPETrainer::readCodes(const char *fp,
+                           unordered_map<tps, uint32_t, pair_hash> &co,
+                           unordered_map<string, tps> &rco) {
   ifstream file(fp);
   if (!file) {
     fprintf(stderr, "Cannot open codes file %s\n", fp);
@@ -419,22 +424,21 @@ void readCodes(const char *fp, unordered_map<tps, uint32_t, pair_hash> &codes,
     assert(splits.size() == 3);
     auto pair = make_pair(splits[0], splits[1]);
     string concat = splits[0] + splits[1];
-    assert(codes.find(pair) == codes.end());
-    assert(reversed_codes.find(concat) == reversed_codes.end());
-    codes[pair] = codes.size();
-    reversed_codes[concat] = pair;
+    assert(co.find(pair) == co.end());
+    assert(rco.find(concat) == rco.end());
+    co[pair] = co.size();
+    rco[concat] = pair;
   }
-  fprintf(stderr, "Read %lu codes from the codes file.\n", codes.size());
+  fprintf(stderr, "Read %lu codes from the codes file.\n", co.size());
 }
 
-void decompose(const string s, vector<string> &newSubwords,
-               const unordered_map<string, tps> &reversed_codes,
-               const unordered_map<string, uint32_t> &vocab, bool isFinal) {
+void BPETrainer::decompose(const string s, vector<string> &newSubwords,
+                           bool isFinal) {
   auto it = reversed_codes.find(s);
   if (it == reversed_codes.end()) {
     // TODO this whole block below is just some sanity check
     // if we cannot un-merge a subword, it has to be a char
-    string s2 = isFinal ? s.substr(0, s.size() - kEndWordLength) : s;
+    string s2 = isFinal ? s.substr(0, s.size() - jEndWordLength) : s;
     int count = 0;
     for (size_t j = 0; j < s2.size(); j++) {
       if ((s2[j] & 0xc0) != 0x80) {
@@ -447,47 +451,43 @@ void decompose(const string s, vector<string> &newSubwords,
   }
   assert(it != reversed_codes.end());
   string token1 = it->second.first;
-  if (vocab.find(token1 + kTokenDelim) == vocab.end()) {
-    decompose(token1, newSubwords, reversed_codes, vocab, false);
+  if (vocab.find(token1 + jTokenDelim) == vocab.end()) {
+    decompose(token1, newSubwords, false);
   } else {
     newSubwords.push_back(token1);
   }
   string token2 = it->second.second;
-  auto query = token2 + kTokenDelim;
+  auto query = token2 + jTokenDelim;
   if (isFinal) {
-    query = token2.substr(0, token2.size() - kEndWordLength);
+    query = token2.substr(0, token2.size() - jEndWordLength);
   }
   if (vocab.find(query) == vocab.end()) {
-    decompose(token2, newSubwords, reversed_codes, vocab, isFinal);
+    decompose(token2, newSubwords, isFinal);
   } else {
     newSubwords.push_back(token2);
   }
 }
 
-void limitVocab(const vector<string> &subwords, vector<string> &newSubwords,
-                const unordered_map<string, tps> &reversed_codes,
-                const unordered_map<string, uint32_t> &vocab) {
+void BPETrainer::limitVocab(const vector<string> &subwords,
+                            vector<string> &newSubwords) {
   string query;
   for (size_t i = 0; i < subwords.size(); i++) {
     bool isFinal = i == subwords.size() - 1;
     auto &subword = subwords[i];
     if (isFinal) {
-      query = subword.substr(0, subword.size() - kEndWordLength);
+      query = subword.substr(0, subword.size() - jEndWordLength);
     } else {
-      query = subword + kTokenDelim;
+      query = subword + jTokenDelim;
     }
     if (vocab.find(query) == vocab.end()) {
-      decompose(subword, newSubwords, reversed_codes, vocab, isFinal);
+      decompose(subword, newSubwords, isFinal);
     } else {
       newSubwords.push_back(subword);
     }
   }
 }
 
-string process_bpe(vector<string> &subwords,
-                   unordered_map<tps, uint32_t, pair_hash> &codes,
-                   unordered_map<string, tps> &reversed_codes,
-                   unordered_map<string, uint32_t> &vocab) {
+string BPETrainer::process_bpe(vector<string> &subwords) {
   // merge subWords as much as possible
   vector<string> newSubwords;
   while (subwords.size() > 1) {
@@ -529,33 +529,21 @@ string process_bpe(vector<string> &subwords,
   // check that we are only using words in the dictionary
   if (vocab.size() > 0) {
     vector<string> newSubwords;
-    limitVocab(subwords, newSubwords, reversed_codes, vocab);
+    limitVocab(subwords, newSubwords);
     subwords = newSubwords;
   }
   // concat subWords
   string result;
   for (auto x : subwords) {
-    result = result + x + kTokenDelim + " ";
+    result = result + x + jTokenDelim + " ";
   }
   return result.substr(
       0,
-      result.size() - kEndWordLength - kTokenDelimLength - 1 // "</w>@@ "
+      result.size() - jEndWordLength - jTokenDelimLength - 1 // "</w>@@ "
   );
 }
 
-void applybpe(const char *outputFile, const char *inputFile,
-              const char *codesPath, const char *vocabPath) {
-  // read vocabulary (to which we want to limit the output file)
-  unordered_map<string, uint32_t> vocab;
-  if (strcmp(vocabPath, "") != 0) {
-    readVocab(vocabPath, vocab);
-  }
-
-  // read codes
-  unordered_map<tps, uint32_t, pair_hash> codes;
-  unordered_map<string, tps> reversed_codes;
-  readCodes(codesPath, codes, reversed_codes);
-
+void BPETrainer::applybpe(const char *outputFile, const char *inputFile) {
   // read input file words
   unordered_map<string, uint32_t> word_count;
   readText(inputFile, word_count);
@@ -570,22 +558,21 @@ void applybpe(const char *outputFile, const char *inputFile,
   }
 
   // apply BPE codes to each word
-  unordered_map<string, string> bpe[kThreads];
+  unordered_map<string, string> bpe[jThreads];
   vector<thread> threads;
-  for (size_t i = 0; i < kThreads; i++) {
+  for (size_t i = 0; i < jThreads; i++) {
     threads.emplace_back(
         [&](size_t this_thread) {
-          for (size_t w = this_thread; w < bpeTokVec.size(); w += kThreads) {
+          for (size_t w = this_thread; w < bpeTokVec.size(); w += jThreads) {
             auto &x = bpeTokVec[w];
-            bpe[this_thread][x.first] =
-                process_bpe(x.second, codes, reversed_codes, vocab);
+            bpe[this_thread][x.first] = process_bpe(x.second);
           }
         },
         i);
   }
 
   unordered_map<string, string> final_bpe;
-  for (size_t i = 0; i < kThreads; i++) {
+  for (size_t i = 0; i < jThreads; i++) {
     threads[i].join();
     for (auto x : bpe[i]) {
       final_bpe[x.first] = x.second;
@@ -595,15 +582,51 @@ void applybpe(const char *outputFile, const char *inputFile,
   outputText(outputFile, inputFile, final_bpe);
 }
 
-BPEApplyer::BPEApplyer(const string &codesPath, const string &vocabPath) {
-  if (vocabPath.size() > 0)
-    readVocab(vocabPath.c_str(), vocab);
-  readCodes(codesPath.c_str(), codes, reversed_codes);
+void BPETrainer::applybpe_stream() {
+  std::string line;
+  while (std::getline(std::cin, line)) {
+    vector<string> tmp;
+    tmp.push_back(line);
+    for (auto &l : apply(tmp)) {
+      std::cout << l << std::endl;
+    }
+  }
 }
 
-vector<string> BPEApplyer::apply(vector<string> &sentences) {
+string BPETrainer::apply(string &sentence) {
+  string cur = "";
+  vector<string> words;
+  split(words, sentence, ' ');
+  for (size_t i = 0; i < words.size(); i++) {
+    auto word = words[i];
+    vector<string> word_bpes;
+    int pos = 0, realLength = 0;
+    int lastStart = 0;
+    while (word[pos]) {
+      bool newChar = (word[pos] & 0xc0) != 0x80; // not a continuation byte
+      realLength += newChar;
+      if (newChar && pos > 0) {
+        auto new_token = word.substr(lastStart, pos - lastStart);
+        word_bpes.push_back(new_token);
+        lastStart = pos;
+      }
+      pos++;
+    }
+    auto bpe = word.substr(lastStart, string::npos) + jEndWord;
+    word_bpes.push_back(bpe);
+    cur += process_bpe(word_bpes);
+    if (i < words.size() - 1)
+      cur += " ";
+  }
+  return cur;
+}
+
+vector<string> BPETrainer::apply(vector<string> &sentences) {
   vector<string> res;
   for (auto &s : sentences) {
+    /*
+    res.emplace_back(apply(s));
+    */
     res.emplace_back("");
     string &cur = res.back();
     vector<string> words;
@@ -623,9 +646,9 @@ vector<string> BPEApplyer::apply(vector<string> &sentences) {
         }
         pos++;
       }
-      auto bpe = word.substr(lastStart, string::npos) + kEndWord;
+      auto bpe = word.substr(lastStart, string::npos) + jEndWord;
       word_bpes.push_back(bpe);
-      cur += process_bpe(word_bpes, codes, reversed_codes, vocab);
+      cur += process_bpe(word_bpes);
       if (i < words.size() - 1)
         cur += " ";
     }
@@ -633,16 +656,24 @@ vector<string> BPEApplyer::apply(vector<string> &sentences) {
   return res;
 }
 
-void applybpe_stream(const char *codesPath, const char *vocabPath) {
-  BPEApplyer applyer(codesPath, vocabPath);
-  std::string line;
-  while (std::getline(std::cin, line)) {
-    vector<string> tmp;
-    tmp.push_back(line);
-    for (auto &l : applyer.apply(tmp)) {
-      std::cout << l << std::endl;
-    }
+BPEInference::BPEInference(const char *codesPath, const char *vocabPath,
+                           const char *jEndWord, const size_t jEndWordLength,
+                           const char *jTokenDelim,
+                           const size_t jTokenDelimLength,
+                           const size_t jThreads)
+    : BPETrainer(jEndWord, jEndWordLength, jTokenDelim, jTokenDelimLength,
+                 jThreads) {
+  unordered_map<string, uint32_t> voc;
+  if (strcmp(vocabPath, "") != 0) {
+    readVocab(vocabPath, voc);
+    vocab = voc;
   }
-}
+
+  unordered_map<tps, uint32_t, pair_hash> co;
+  unordered_map<string, tps> rco;
+  readCodes(codesPath, co, rco);
+  codes = co;
+  reversed_codes = rco;
+};
 
 } // namespace flexBPE
